refactor @ConfigurationLocations / @PropertySource
Use pure ConfigurationClassParser PropertySource

We load all named PropertySources in the choosen order with annotation on
the first loaded Configuration class. First declared PropertySource is
the one with the minimal precedence. Initializer only handle adding of
property from bootstrap*properties files.

So we have :

* Initializer that loads bootstrap files, so it allows to control
  profiles and igloo.configurationLocations setting, according on
  potential system or environment properties. Properties files from
  the igloo.configurationLocations have the higher precedence.
* First configuration class loads, with @PropertySource annotation
  named property sources for component, framework and application
  levels. component and framework levels are empty.
* @PropertySource must be used with the appropriate name to override
  only desired level. From higher to lower precedende : application,
  framework, component. Last parsed PropertySource for a given level is
  the one with the highest precedence.
* WIP: we need a better way to order PropertySource-s

WIP, as the problem is that configuration.properties is not loaded with
a highest precedence than autoconfigure @PropertySource'd files, we may
choose to trigger loading of configuration.properties by initializer
with igloo.configurationLocations property.

This solution needs to remove
@PropertySource("classpath:/configuration.properties") from
@Configuration beans and is not backward-compatible as we modify
configuration-bootstrap-default.properties.

Removed jpa-common.properties from autoconfiguration as
@PropertySource from autoconfiguration as @PropertySource are loaded
in the order indicated by @Order on @Configuration bean, BUT
autoconfigure beans are always loaded after "classic" beans,
jpa-common.properties was added with an higher precedence.