jpa-split: refactor module batch-executor
Changes:

batch-executor is pushed in -api, [] and -integration modules.
This modules are optional for applications.

Autoconfiguration is modified to conditionally build related beans. It
is linked to classpath detection and
igloo-ac.batch-executor.disabled property.

igloo-component-jpa-batch-executor-api allows third-party module to
implement a IBatchExecutorListener. Implementation are provided for
hibernate-search (to trigger flushToIndexes and reindexEntities) and
task management (to trigger beforeClear() / it performs beforeCommit
tasks it there are tagged as shouldRunBeforeClear).

Some errors in loggers and code-style violations are fixed in
batch-executor.

Some tests are added in autoconfigure module.

Some fixes on spring context management are related to the previous
commit.

Created modules:
* igloo-component-jpa-batch-executor
* igloo-component-jpa-batch-executor-api
* igloo-component-jpa-batch-executor-integration

Incompatible changes:
* org.iglooproject.jpa.batch.util.IBeforeClearListener moved to
  org.iglooproject.jpa.batch.api.IBeforeClearListener and is deprecated.
  Uses org.iglooproject.jpa.batch.api.IBatchExecutorListener.
* batch executor Spring configuration is no longer loaded by default for
  old-style (pre-autoconfigure) configuration. JpaBatchExecutorConfig
  and HibernateSearchBatchExecutorConfig must be imported if needed
  (for new-style configuration, classpath detection and
  igloo-ac.*.disabled are used to autoconfigure spring context)